Explicitly use write-without-response flag
Advertise the explicit use of writing WITHOUT response.

This _could_ fix an issue relating to some devices checking for write
type before writing, and then failing the write due to reading it as
writable WITH response.

In any case, it's good to make the characteristic actually reflect what
is happening, even if it doesn't result in any significant change in
performance.

Signed-off-by: Elijah J. Passmore <elijahjpassmore@elijahjpassmore.com>